requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pypdf>=3.17.0
matplotlib>=3.8.0
pillow>=10.1.0
requests>=2.31.0